// src/app/common/models/tableReservation.ts
export class TableReservation {
    _id: String;
    FirstName: String;
    LastName:String;
    TableId:String;
    UserId:String;
    ReservationCode:String;
    RestaurantId: String;
    RestaurantName: String;
    TableReservationQuantity: Number;
    TableReservationDate: Date;
    CreatedDate: Date;
    ModifiedDate: Date;
}